pyproject.toml
[tool.poetry]
name = "pyrest"
version = "0.1.0"
description = ""
authors = ["Chart <chart.asa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
certifi = "2021.10.8"
charset-normalizer = "2.0.12"
click = "8.1.3"
Flask = "2.1.2"
Flask-Login = "0.6.1"
Flask-SQLAlchemy = "2.5.1"
greenlet = "1.1.2"
idna = "3.3"
importlib-metadata = "4.11.3"
itsdangerous = "2.1.2"
Jinja2 = "3.1.2"
MarkupSafe = "2.1.1"
requests = "2.27.1"
SQLAlchemy = "1.4.36"
urllib3 = "1.26.9"
Werkzeug = "2.1.2"
zipp = "3.8.0"
python-dotenv = "^0.20.0"
SQLAlchemy-Utils = "^0.38.2"
psycopg2-binary = "^2.9.3"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
